x/socialapp/types/message_update_profile_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"socialapp/testutil/sample"
)

func TestMsgUpdateProfile_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateProfile
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateProfile{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateProfile{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
